feat(apiEndPoint): add POST api endpoint to create new red flag
user should be able to create new red-flag record

[Delivers #162393207]
