--- server/controllers/red-flags.js
@@ -1,9 +1,0 @@
-import getRedFlags from './getRedFlags';
-import getRedFlag from './getRedFlag';
-import createRedFlag from './createRedFlag';
-
-export default {
-	getRedFlags,
-	getRedFlag,
-	createRedFlag
-}
