+++ server/routes/red-flags.js
@@ -1,12 +1,13 @@
 import express from 'express';
+import routes from '../controllers/redFlags';
 import controllers from '../controllers/red-flags.js';
 
 const router = express.Router();
 
-router.get('/red-flags', controllers.getRedFlags);
+router.get('/red-flags', routes.getRedFlags);
 
-router.post('/red-flags', controllers.createRedFlag);
+router.post('/red-flags', routes.createRedFlag);
 
-router.get('/red-flags/:id', controllers.getRedFlag);
+router.get('/red-flags/:id', routes.getRedFlag);
 
 export default router;
